requirements.txt
fastapi
uvicorn
jinja2
boto3
gitpython
